Last stable version 4 on defense

+++ server/src/main/java/com/edu_netcracker/cmp/RestControllers/UsersController.java
@@ -5,6 +5,7 @@ import com.edu_netcracker.cmp.entities.DTO.StudentsAttributesDTO;
 import com.edu_netcracker.cmp.entities.DTO.TemplateDto;
 import com.edu_netcracker.cmp.entities.DTO.UserDTO;
 import com.edu_netcracker.cmp.entities.StudentsToAttributes;
+import com.edu_netcracker.cmp.entities.jpa.AttributesJPA;
 import com.edu_netcracker.cmp.entities.jpa.STAJPA;
 import com.edu_netcracker.cmp.entities.jpa.UsersJpa;
 import com.edu_netcracker.cmp.usersInfoService.UserInfoService;
@@ -12,6 +13,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.w3c.dom.Attr;
 
@@ -28,6 +32,8 @@ public class UsersController {
 
     private UserInfoService userInfoService;
     private UsersJpa usersJpa;
+    @Autowired
+    private AttributesJPA attributesJPA;
 
     @Autowired
     public UsersController(UserInfoService userInfoService, UsersJpa usersJpa) {
@@ -43,17 +49,30 @@ public class UsersController {
 
     @PostMapping("saveForUser/{userId}")
     public ResponseEntity<Void> testSave(@RequestBody Map<String, Object> attributes, @PathVariable String userId) {
+        UserDetails userDetails;
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        userDetails = (UserDetails) authentication.getPrincipal();
 
-        userInfoService.modifyAttributes(attributes, userId);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (userDetails.getAuthorities().stream().anyMatch(x-> x.getAuthority().equals("admin")) || userDetails.getUsername().equals(userId)) {
+            userInfoService.modifyAttributes(attributes, userId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     @PostMapping("/delete/{userId}")
     public ResponseEntity<Void> deleteAttribute(@RequestBody String attributeName, @PathVariable String userId){
-        Attributes attributes = new Attributes();
-        attributes.setAttributeName(attributeName);
-        userInfoService.deleteAttribute(attributes, userId);
-        return new ResponseEntity<>(HttpStatus.OK);
+        UserDetails userDetails;
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        userDetails = (UserDetails) authentication.getPrincipal();
+
+        if (userDetails.getAuthorities().stream().anyMatch(x-> x.getAuthority().equals("admin"))
+                || userDetails.getUsername().equals(userId)) {
+            Attributes attributes = attributesJPA.findAttributesByAttributeName(attributeName);
+            userInfoService.deleteAttribute(attributes, userId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     @GetMapping("getAll")
@@ -62,10 +81,32 @@ public class UsersController {
         return new ResponseEntity<>(usersByUserNameDTO, HttpStatus.OK);
     }
 
+    @PostMapping("createUser")
+    public ResponseEntity<Void> createUser(@RequestBody UserCreationDto creationDto) {
+        UserDetails userDetails;
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        userDetails = (UserDetails) authentication.getPrincipal();
+
+        if (userDetails.getAuthorities().stream().anyMatch(x-> x.getAuthority().equals("admin")) ){
+            userInfoService.createUser(creationDto);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+    }
+
     @PostMapping("sendToUser/{userId}")
     public ResponseEntity<Void> sendToUser(@RequestBody TemplateDto tempalte, @PathVariable String userId) {
-        userInfoService.sendDirectlyToUser(tempalte, userId);
-        return  new ResponseEntity(HttpStatus.OK);
+        UserDetails userDetails;
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        userDetails = (UserDetails) authentication.getPrincipal();
+
+        if (userDetails.getAuthorities().stream().anyMatch(x-> x.getAuthority().equals("admin"))
+                || userDetails.getAuthorities().stream().anyMatch(x-> x.getAuthority().equals("hr"))
+                ) {
+            userInfoService.sendDirectlyToUser(tempalte, userId);
+            return new ResponseEntity(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
